requirements.txt
google-cloud-firestore
tensorflow
pokebase
numpy
firebase-admin